# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.AdminCreateUserRequest:
    properties:
      avatar:
        maxLength: 255
        type: string
      badge:
        maxLength: 50
        type: string
      email:
        type: string
      nickname:
        maxLength: 32
        minLength: 2
        type: string
      password:
        maxLength: 32
        minLength: 6
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.UserRole'
        enum:
        - super_admin
        - admin
        - user
        - guest
      website:
        maxLength: 255
        type: string
    required:
    - email
    - nickname
    - password
    - role
    type: object
  dto.AdminUpdateUserRequest:
    properties:
      avatar:
        maxLength: 255
        type: string
      badge:
        maxLength: 50
        type: string
      email:
        type: string
      is_enabled:
        description: 使用指针以区分未设置和false
        type: boolean
      nickname:
        maxLength: 32
        minLength: 2
        type: string
      password:
        description: 可选的密码字段，用于更新密码
        maxLength: 32
        minLength: 6
        type: string
      role:
        allOf:
        - $ref: '#/definitions/model.UserRole'
        enum:
        - super_admin
        - admin
        - user
        - guest
      website:
        maxLength: 255
        type: string
    type: object
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/dto.UserResponse'
        description: 刷新token时为空
    type: object
...
  dto.UserResponse:
    properties:
      avatar:
        type: string
      badge:
        type: string
      created_at:
        $ref: '#/definitions/utils.JSONTime'
      email:
        type: string
      email_hash:
        description: 邮箱哈希（用于 Gravatar）
        type: string
      has_password:
        type: boolean
      id:
        type: integer
      is_enabled:
        type: boolean
      is_virtual_email:
        description: 是否为虚拟邮箱（需绑定真实邮箱）
        type: boolean
      last_login:
        $ref: '#/definitions/utils.JSONTime'
      linked_oauths:
        description: '["github", "google", "qq"]'
        items:
          type: string
        type: array
      nickname:
        type: string
      role:
        $ref: '#/definitions/model.UserRole'
      website:
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: talen2004@163.com
    name: M.Talen
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Flec-Server
  version: v1
paths:
...
  /admin/users:
    get:
      consumes:
      - application/json
      description: 获取所有用户
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 用户列表
      tags:
      - 用户管理
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminCreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建用户
      tags:
      - 用户管理
  /admin/users/{id}:
    delete:
      consumes:
      - application/json
      description: 软删除用户
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除用户
      tags:
      - 用户管理
    get:
      consumes:
      - application/json
      description: 查看用户详细信息
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 用户详情
      tags:
      - 用户管理
    put:
      consumes:
      - application/json
      description: 修改用户信息
      parameters:
      - description: 用户 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息（包含可选的密码字段）
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AdminUpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户
      tags:
      - 用户管理
...
  /auth/{provider}:
    get:
      description: 跳转到第三方认证页面，支持登录和绑定
      parameters:
      - description: 提供商 (github/google/qq)
        in: path
        name: provider
        required: true
        type: string
      - description: 操作类型：login(默认) 或 bind
        in: query
        name: action
        type: string
      - description: 绑定时的认证token
        in: query
        name: token
        type: string
      - description: 成功后跳转的页面路径
        in: query
        name: redirect
        type: string
      responses: {}
      summary: 第三方认证
      tags:
      - 认证
  /auth/{provider}/callback:
    get:
      description: 处理第三方登录回调，返回 Token 并跳转
      parameters:
      - description: 提供商 (github/google/qq)
        in: path
        name: provider
        required: true
        type: string
      responses: {}
      summary: 第三方认证回调
      tags:
      - 认证
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: 通过邮箱找回密码，发送重置验证码
      parameters:
      - description: 邮箱地址
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 重置密码请求
      tags:
      - 认证
  /auth/login:
    post:
      consumes:
      - application/json
      description: 邮箱密码登录，返回 JWT token 和用户基本信息
      parameters:
      - description: 登录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户登录
      tags:
      - 认证
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 将当前 token 加入黑名单使其失效
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用refresh token获取新的access token
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: 刷新token
      tags:
      - 认证
  /auth/register:
    post:
      consumes:
      - application/json
      description: 邮箱密码注册
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      summary: 用户注册
      tags:
      - 认证
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: 凭验证码设置新密码，成功后需重新登录
      parameters:
      - description: 重置密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: 重置密码
      tags:
      - 认证
...
  /user/deactivate:
    delete:
      consumes:
      - application/json
      description: 主动注销自己的账号，需提供密码验证，注销后账号将被软删除
      parameters:
      - description: 注销账号信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeactivateAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 注销账号
      tags:
      - 用户
  /user/oauth/{provider}:
    delete:
      consumes:
      - application/json
      description: 解绑已绑定的第三方账号
      parameters:
      - description: 提供商 (github/google)
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 解绑第三方账号
      tags:
      - 用户
  /user/password:
    post:
      consumes:
      - application/json
      description: OAuth 注册用户首次设置密码，无需旧密码验证
      parameters:
      - description: 密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 设置密码
      tags:
      - 用户
    put:
      consumes:
      - application/json
      description: 修改密码，需提供旧密码验证
      parameters:
      - description: 密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 修改密码
      tags:
      - 用户
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"